a function that returns the length of a string.

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -2,13 +2,13 @@
 
 /**
  * _strlen - function that returns the length of a string.
- * @s: returns the length of a string
+ * @s: parameter 
  * Return: length of a string
  */
 
 int _strlen(char *s)
 {
-	size_t length = 0;
+	int length = 0;
 
 	while (*str++)
 		length++;
